move OurValues into its own compo
