+++ app/about_us/OurValues.js
@@ -1,0 +1,44 @@
+import styles from "./OurValues.module.css";
+
+const OurValues = () => {
+  return (
+    <>
+      <h1 className={styles.hugeH1}>01</h1>
+      <h2>Exploration</h2>
+      <p>
+        We are driven by a deep-seated desire to explore the unknown. We believe
+        that the pursuit of discovery is at the heart of human nature, and we
+        are committed to pushing the boundaries of what is possible.
+      </p>
+      <div></div>
+      <h1>02</h1>
+      <h2>Innovation</h2>
+      <p>
+        At Galactica, we prioritize cutting-edge technology and innovation. We
+        are constantly evolving our spacecraft, safety protocols, and services
+        to ensure that our travelers experience the most advanced and secure
+        space journeys available.
+      </p>
+      <div></div>
+      <h1>03</h1>
+      <h2>Sustainability</h2>
+      <p>
+        We are committed to making space exploration sustainable for future
+        generations. Our space missions are designed to minimize environmental
+        impact, both on Earth and in space, and to foster a spirit of
+        responsibility towards our universe.
+      </p>
+      <div></div>
+      <h1>04</h1>
+      <h2>Community</h2>
+      <p>
+        We believe in the power of collective exploration. Our journeys are not
+        just about reaching new destinations; they are about building a
+        community of space enthusiasts who share a passion for the stars.
+      </p>
+      <div></div>
+    </>
+  );
+};
+
+export default OurValues;
