+++ app/about_us/OurValues.module.css
@@ -1,0 +1,3 @@
+.hugeH1 {
+  font-size: 4.5rem;
+}
